+++ familyhelper-finance-impl/src/main/java/com/dwarfeng/familyhelper/finance/impl/handler/AccountBookOperateHandlerImpl.java
@@ -17,8 +17,11 @@ import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import org.springframework.stereotype.Component;
 
 import java.math.BigDecimal;
+import java.util.Collection;
 import java.util.Date;
+import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
 @Component
 public class AccountBookOperateHandlerImpl implements AccountBookOperateHandler {
@@ -121,7 +124,7 @@ public class AccountBookOperateHandlerImpl implements AccountBookOperateHandler 
     }
 
     @Override
-    public void addGuestPermission(StringIdKey ownerUserKey, StringIdKey guestUserKey, LongIdKey accountBookKey)
+    public void addGuestPermission(StringIdKey ownerUserKey, LongIdKey accountBookKey, StringIdKey guestUserKey)
             throws HandlerException {
         try {
             // 1. 如果用户主键与访客主键一致，则什么也不做。
@@ -154,7 +157,7 @@ public class AccountBookOperateHandlerImpl implements AccountBookOperateHandler 
     }
 
     @Override
-    public void removeGuestPermission(StringIdKey ownerUserKey, StringIdKey guestUserKey, LongIdKey accountBookKey)
+    public void removeGuestPermission(StringIdKey ownerUserKey, LongIdKey accountBookKey, StringIdKey guestUserKey)
             throws HandlerException {
         try {
             // 1. 如果用户主键与访客主键一致，则什么也不做。
@@ -175,6 +178,51 @@ public class AccountBookOperateHandlerImpl implements AccountBookOperateHandler 
             // 5. 通过入口信息组合权限实体主键，并进行存在删除操作。
             PoabKey poabKey = new PoabKey(accountBookKey.getLongId(), guestUserKey.getStringId());
             poabMaintainService.deleteIfExists(poabKey);
+        } catch (HandlerException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new HandlerException(e);
+        }
+    }
+
+    @Override
+    public void resetGuestPermission(
+            StringIdKey ownerUserKey, LongIdKey accountBookKey, Collection<StringIdKey> guestUserKeys
+    ) throws HandlerException {
+        try {
+            // 1. 确认用户存在。
+            makeSureUserExists(ownerUserKey);
+            for (StringIdKey guestUserKey : guestUserKeys) {
+                makeSureUserExists(guestUserKey);
+            }
+
+            // 2. 确认账本存在。
+            makeSureAccountBookExists(accountBookKey);
+
+            // 3. 确认用户有权限操作指定的账本。
+            makeSureUserPermittedForAccountBook(ownerUserKey, accountBookKey);
+
+            // 4. 查询并删除个人简介对应的所有权限。
+            List<PoabKey> poabKeys = poabMaintainService.lookup(
+                    PoabMaintainService.CHILD_FOR_ACCOUNT_BOOK_PERMISSION_LEVEL_EQUALS,
+                    new Object[]{accountBookKey, Poab.PERMISSION_LEVEL_GUEST}
+            ).getData().stream().map(Poab::getKey).collect(Collectors.toList());
+            poabMaintainService.batchDeleteIfExists(poabKeys);
+
+            // 4. 遍历集合，组合权限实体列表。
+            List<Poab> poabs = guestUserKeys.stream().distinct().filter(
+                    key -> !Objects.equals(ownerUserKey, key)
+            ).map(
+                    key -> new Poab(
+                            new PoabKey(accountBookKey.getLongId(), key.getStringId()),
+                            Poab.PERMISSION_LEVEL_GUEST,
+                            "赋予用户 " + key.getStringId() + " 访客权限"
+                    )
+            ).collect(Collectors.toList());
+            if (poabs.isEmpty()) {
+                return;
+            }
+            poabMaintainService.batchInsert(poabs);
         } catch (HandlerException e) {
             throw e;
         } catch (Exception e) {