+++ familyhelper-finance-stack/src/main/java/com/dwarfeng/familyhelper/finance/stack/handler/AccountBookOperateHandler.java
@@ -7,6 +7,8 @@ import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.HandlerException;
 import com.dwarfeng.subgrade.stack.handler.Handler;
 
+import java.util.Collection;
+
 /**
  * 账本处理器。
  *
@@ -51,21 +53,32 @@ public interface AccountBookOperateHandler extends Handler {
      * 添加账本的访客权限。
      *
      * @param ownerUserKey   账本的所有者的主键。
-     * @param guestUserKey   访客的主键。
      * @param accountBookKey 账本的主键
+     * @param guestUserKey   访客的主键。
      * @throws HandlerException 处理器异常。
      */
-    void addGuestPermission(StringIdKey ownerUserKey, StringIdKey guestUserKey, LongIdKey accountBookKey)
+    void addGuestPermission(StringIdKey ownerUserKey, LongIdKey accountBookKey, StringIdKey guestUserKey)
             throws HandlerException;
 
     /**
      * 移除账本的访客权限。
      *
      * @param ownerUserKey   账本的所有者的主键。
-     * @param guestUserKey   访客的主键。
      * @param accountBookKey 账本的主键
+     * @param guestUserKey   访客的主键。
      * @throws HandlerException 处理器异常。
      */
-    void removeGuestPermission(StringIdKey ownerUserKey, StringIdKey guestUserKey, LongIdKey accountBookKey)
+    void removeGuestPermission(StringIdKey ownerUserKey, LongIdKey accountBookKey, StringIdKey guestUserKey)
+            throws HandlerException;
+
+    /**
+     * 重置账本的访客权限。
+     *
+     * @param ownerUserKey   账本的所有者的主键。
+     * @param accountBookKey 账本的主键
+     * @param guestUserKeys  访客的主键组成的集合。
+     * @throws HandlerException 处理器异常。
+     */
+    void resetGuestPermission(StringIdKey ownerUserKey, LongIdKey accountBookKey, Collection<StringIdKey> guestUserKeys)
             throws HandlerException;
 }