+++ familyhelper-finance-stack/src/main/java/com/dwarfeng/familyhelper/finance/stack/service/AccountBookOperateService.java
@@ -7,6 +7,8 @@ import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import com.dwarfeng.subgrade.stack.service.Service;
 
+import java.util.Collection;
+
 /**
  * 账本操作服务。
  *
@@ -51,21 +53,32 @@ public interface AccountBookOperateService extends Service {
      * 添加账本的访客权限。
      *
      * @param ownerUserKey   账本的所有者的主键。
-     * @param guestUserKey   访客的主键。
      * @param accountBookKey 账本的主键
+     * @param guestUserKey   访客的主键。
      * @throws ServiceException 服务异常。
      */
-    void addGuestPermission(StringIdKey ownerUserKey, StringIdKey guestUserKey, LongIdKey accountBookKey)
+    void addGuestPermission(StringIdKey ownerUserKey, LongIdKey accountBookKey, StringIdKey guestUserKey)
             throws ServiceException;
 
     /**
      * 移除账本的访客权限。
      *
      * @param ownerUserKey   账本的所有者的主键。
-     * @param guestUserKey   访客的主键。
      * @param accountBookKey 账本的主键
+     * @param guestUserKey   访客的主键。
      * @throws ServiceException 服务异常。
      */
-    void removeGuestPermission(StringIdKey ownerUserKey, StringIdKey guestUserKey, LongIdKey accountBookKey)
+    void removeGuestPermission(StringIdKey ownerUserKey, LongIdKey accountBookKey, StringIdKey guestUserKey)
+            throws ServiceException;
+
+    /**
+     * 重置账本的访客权限。
+     *
+     * @param ownerUserKey   账本的所有者的主键。
+     * @param accountBookKey 账本的主键
+     * @param guestUserKeys  访客的主键组成的集合。
+     * @throws ServiceException 服务异常。
+     */
+    void resetGuestPermission(StringIdKey ownerUserKey, LongIdKey accountBookKey, Collection<StringIdKey> guestUserKeys)
             throws ServiceException;
 }